WinAdapter: Export symbols for CoTaskMem{Alloc,Free}
Some functions return lists allocated with CoTaskMemAlloc.  Applications
using libdxcompiler.so currently have to look into the code to
understand that these should be deallocated with free() instead of
delete[].  Export the CoTaskMemFree function so that there is no
guesswork involved anymore: the application can now call this function
and always free it in the right way.